extracts testdata to own file
fixes decimals validation
