+++ src/__tests__/transfers.test.ts
@@ -5,6 +5,7 @@ import { TokenInfo } from "@uniswap/token-lists";
 import BigNumber from "bignumber.js";
 import { fetchTokenList, TokenMap } from "src/hooks/tokenList";
 import { Payment } from "src/parser";
+import { testData } from "../test/util";
 
 let dummySafeInfo: SafeInfo = {
   safeAddress: "0x123",
@@ -15,20 +16,13 @@ let dummySafeInfo: SafeInfo = {
 let tokenList: TokenMap;
 let listedTokens: string[];
 let listedToken: TokenInfo;
-const receiverAddress = "0x1000000000000000000000000000000000000000";
-const unlistedToken: TokenInfo = {
-  address: "0x6b175474e89094c44da98b954eedeac495271d0f",
-  decimals: 18,
-  symbol: "UNL",
-  name: "Unlisted",
-  chainId: -1,
-};
+const receiverAddress = testData.addresses.receiver1;
 
 // TODO - make method erc20TransferData and replace data checks with this instead of hardcoded strings.
 // function erc20TransferData(amount: BigNumber, receiver: string): Bytes {}
 
 describe("Build Transfers:", () => {
-  beforeEach(async () => {
+  beforeAll(async () => {
     tokenList = await fetchTokenList(dummySafeInfo.network);
     listedTokens = Array.from(tokenList.keys());
     listedToken = tokenList.get(listedTokens[0]);
@@ -50,8 +44,8 @@ describe("Build Transfers:", () => {
       {
         receiver,
         amount: amount,
-        tokenAddress: unlistedToken.address,
-        decimals: unlistedToken.decimals,
+        tokenAddress: testData.unlistedToken.address,
+        decimals: testData.unlistedToken.decimals,
       },
       // Native Asset
       {
@@ -85,7 +79,7 @@ describe("Build Transfers:", () => {
       );
 
       expect(unlistedTransfer.value).to.be.equal("0");
-      expect(unlistedTransfer.to).to.be.equal(unlistedToken.address);
+      expect(unlistedTransfer.to).to.be.equal(testData.unlistedToken.address);
       expect(unlistedTransfer.data).to.be.equal(
         "0xa9059cbb0000000000000000000000001000000000000000000000000000000000000000800000000000016c889a28c160ce0422bb9138ff1d4e48274000000000000000"
       );
@@ -117,7 +111,7 @@ describe("Build Transfers:", () => {
       );
 
       expect(unlisted.value).to.be.equal("0");
-      expect(unlisted.to).to.be.equal(unlistedToken.address);
+      expect(unlisted.to).to.be.equal(testData.unlistedToken.address);
       expect(unlisted.data).to.be.equal(
         "0xa9059cbb0000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000174876e800"
       );
@@ -146,7 +140,7 @@ describe("Build Transfers:", () => {
       );
 
       expect(unlisted.value).to.be.equal("0");
-      expect(unlisted.to).to.be.equal(unlistedToken.address);
+      expect(unlisted.to).to.be.equal(testData.unlistedToken.address);
       expect(unlisted.data).to.be.equal(
         "0xa9059cbb0000000000000000000000001000000000000000000000000000000000000000000000000000000000000000000000000000000000001a24902beecbd5109200"
       );
